# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Upakovali online API documentation",
        "contact": {
            "email": "ib@mitlabs.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://upakovali-online.herokuapp.com/api/v1",
            "description": "Тестовый url для методов API (префикс)"
        },
        {
            "url": "http://localhost:29080/api/v1",
            "description": "Локальный url для методов API (префикс)"
        }
    ],
    "paths": {
        "/login/parthner": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Получить токен доступа",
                "operationId": "loginParthner",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parthner_hash": {
                                        "description": "Parthner hash",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "parthner_hash": "$2y$10$XVxaSMF64H3NdAUfkULvFuLESv1w5GaokWUJMzpQgAR/bY8zsCa6a"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            }
        },
        "/outer-api/updateOrder": {
            "patch": {
                "tags": [
                    "Order"
                ],
                "summary": "Изменить время приема/доставки заказа",
                "operationId": "updateOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "external_number": {
                                        "description": "Внешний номер заказа",
                                        "type": "string"
                                    },
                                    "receiving_date": {
                                        "description": "Время приема в формате (d-m-Y H:i)",
                                        "type": "string"
                                    },
                                    "issue_date": {
                                        "description": "Время выдачи в формате (d-m-Y H:i)",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "external_number": "04073d28-5ffd-3715-b7fa-77736767bf29",
                                    "receiving_date": "09-11-2021 15:21",
                                    "issue_date": "09-11-2021 16:00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/outer-api/setStatus": {
            "patch": {
                "tags": [
                    "Order"
                ],
                "summary": "Изменить статус заказа",
                "operationId": "setStatus",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "external_number": {
                                        "description": "Внешний номер заказа",
                                        "type": "string"
                                    },
                                    "order_status_id": {
                                        "description": "Новый статус заказа ( 1 - Создан, 2 - Курьер назначен, 3 - Забрали, 4 - Приняли, 5 - В работе, 6 - Упаковали, 7 - Выдали, 8 - Курьер выдает, 9 - Доставили, 10 - Оплатили, 11 - Отмена)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "external_number": "04073d28-5ffd-3715-b7fa-77736767bf29",
                                    "order_status_id": 2
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/outer-api/cancelOrder": {
            "patch": {
                "tags": [
                    "Order"
                ],
                "summary": "Отменить заказ",
                "operationId": "cancelOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "external_number": {
                                        "description": "Внешний номер заказа",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "external_number": "04073d28-5ffd-3715-b7fa-77736767bf29"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "title": "Client",
                "description": "Клиент",
                "properties": {
                    "id": {
                        "title": "Id",
                        "description": "Id клиента",
                        "format": "integer",
                        "example": 1
                    },
                    "full_name": {
                        "title": "Full name",
                        "description": "ФИО",
                        "type": "string",
                        "format": "string",
                        "example": "Любимов Алексей Валерьевич"
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "Телефон",
                        "type": "string",
                        "format": "string",
                        "example": "+79586526652"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "format": "email",
                        "example": "sergeev2@example.com"
                    },
                    "comment": {
                        "title": "Comment",
                        "description": "Комментарий",
                        "type": "string",
                        "format": "string",
                        "example": "Лучше купи себе часы"
                    }
                },
                "type": "object"
            },
            "Parthner": {
                "title": "Parthner",
                "description": "Партнер",
                "properties": {
                    "id": {
                        "title": "Id",
                        "description": "Id партнера",
                        "format": "integer",
                        "example": 1
                    },
                    "full_name": {
                        "title": "Full name",
                        "description": "ФИО",
                        "type": "string",
                        "format": "string",
                        "example": "Любимов Алексей Валерьевич"
                    },
                    "manager": {
                        "$ref": "#/components/schemas/User"
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "Телефон",
                        "type": "string",
                        "format": "string",
                        "example": "+79586526652"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "format": "email",
                        "example": "sergeev2@example.com"
                    },
                    "comment": {
                        "title": "Comment",
                        "description": "Комментарий",
                        "type": "string",
                        "format": "string",
                        "example": "Лучше купи себе часы"
                    }
                },
                "type": "object"
            },
            "Role": {
                "title": "Role",
                "description": "Роль пользователя",
                "properties": {
                    "id": {
                        "title": "Id",
                        "description": "Id роли",
                        "format": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Название роли",
                        "type": "string",
                        "format": "string",
                        "example": "Мастер"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Пользователь",
                "properties": {
                    "id": {
                        "title": "Id",
                        "description": "Id пользователя",
                        "format": "integer",
                        "example": 1
                    },
                    "full_name": {
                        "title": "Full name",
                        "description": "ФИО",
                        "type": "string",
                        "format": "string",
                        "example": "Любимов Алексей Валерьевич"
                    },
                    "manager": {
                        "title": "manager",
                        "description": "Менеджер",
                        "format": "string"
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "Телефон",
                        "type": "string",
                        "format": "string",
                        "example": "+79586526652"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "format": "email",
                        "example": "sergeev2@example.com"
                    },
                    "comment": {
                        "title": "Comment",
                        "description": "Комментарий",
                        "type": "string",
                        "format": "string",
                        "example": "Лучше купи себе часы"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer Token": {
                "type": "http",
                "description": "Login with parthner_hash to get the authentication token",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}